backend/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.2
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
